+++ db/migrate/20220819121534_add_not_applicable_to_task.rb
@@ -1,0 +1,5 @@
+class AddNotApplicableToTask < ActiveRecord::Migration[7.0]
+  def change
+    add_column :tasks, :not_applicable, :boolean, default: false
+  end
+end