+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_08_18_142154) do
+ActiveRecord::Schema[7.0].define(version: 2022_08_19_121534) do
   create_table "actions", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.string "title", null: false
     t.integer "order", null: false
@@ -79,6 +79,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_08_18_142154) do
     t.text "hint"
     t.string "guidance_summary"
     t.text "guidance_text"
+    t.boolean "optional"
+    t.boolean "not_applicable", default: false
     t.index ["section_id"], name: "index_tasks_on_section_id"
   end