+++ spec/factories/task_factory.rb
@@ -14,5 +14,16 @@ FactoryBot.define do
     order { 0 }
 
     section
+    optional { false }
+    not_applicable { false }
+
+    trait :optional do
+      optional { true }
+    end
+
+    trait :not_applicable do
+      optional { true }
+      not_applicable { true }
+    end
   end
 end