+++ CHANGELOG.md
@@ -21,9 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Project contacts can be added and edited
 - Add hint text and guidance text to actions.
 - A regional delivery officer must be selected at the time of project creation.
-  A regional delivery officer can only see project they are assigned to on the
+- A regional delivery officer can only see project they are assigned to on the
   projects index page.
 - Add hint text and guidance text to tasks.
+- Optional tasks can be marked 'Not applicable' to the project
 
 ### Removed